// src/app/models/info.ts
export interface Information {
    id: number;
    title: string;
    object: string;
    datPost: Date;
    destinataire: number;
    ref: string;
    datEven: Date;
    description: string;
    periode: Date;
    file: File;
    admin_id: number;
    niveauEtudes: number[];
    specialites: string[];
}